+++ glade_forms/CHANGELOG.md
@@ -5,6 +5,8 @@
 - **[Add]**: Add `GladeComposedFormConsumer`
 - **[Add]**: Add `GladeComposedModelProvider` to provide `GladeComposedModel`
 - **[Add]**: Add `ComposedExample` to demonstrate `GladeComposedModel` functionality
+- **[Add]**: Add `GladeModel.fillDebugMetadata()` method to provide debug metadata as key-value pairs.
+  - This metadata is displayed in `GladeModelDebugInfo` widget.
 
 ## 5.0.0